Classes are becoming a mess. Switching to utility classes.
This is experimental and I want to see how it goes. My classes were becoming very messy... probably because I don't have a prototype for this website to go off of and I was designing as I developed. So I kept finding the need to have to change my classes to accommodate something new but this is problematic because then it affects other parts of the layout that I wanted to leave as-is. I think if I use utility classes explicitly throughout the layout then it may be easier to manipulate different sections this way and possibly easier to maintain over time. It's tedious to make this switch but I think it'll pay off...