WordPress 6.4.1 update conflict with reference to scripts.js
WordPress 6.4.1 did not like how I had functions.php configured to call for scripts.js with the 'type' attribute set to 'module'. It was setting the 'type' attribute to 'text/javascript' instead. This is problematic because we need it set to 'module' so that we can import glide.js within it.

I'm not exactly sure why it didn't like what I had but it is now rewritten to simply generate the entire line <script> line of code rather than use str_replace to swap out the type value.